[7418] Create common util processPopulateExpression for DateTime and Time controls
